Support for Huasifei WH3000 (Fudy MT3000)
Everything is working on a recent snapshot except it formats mmcblck0p6 to ~500MB instead of 7.3GB. Maybe I've missed some settings or packages for the device?

Also there eth0 LAN is not working but it has proper MAC. And system doesn't save settings.

Need your help to fix it. I have UART.