build: optionally provide pigz through a symlink
pigz[1] is a drop-in/fully functional replacement for gzip that exploits
multiple processors and multiple cores when compressing data. If pigz is
present on the host system, create the symlink for it under
staging_dir/host/bin/ by calling SetupHostCommand.

If pigz is not on the host, legacy gzip will be used.

1. https://www.zlib.net/pigz

Signed-off-by: John Audia <therealgraysky@proton.me>